completed functionality for publisher and subscriber endpoints
