+++ package.json
@@ -21,7 +21,6 @@
   },
   "homepage": "https://github.com/olorondu-emeka/http-notification-system#readme",
   "dependencies": {
-    "axios": "^0.21.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-ws": "^4.0.0",
